fix image attach bug for reward

+++ app/Http/Controllers/Api/ParentSide/RewardController.php
@@ -69,6 +69,7 @@ class RewardController extends BaseController
             
             return $this->sendResponseWithData($reward->load(['image']), 200);
         } else {
+            
             return $this->sendResponseWithData($reward->withoutRelations(), 200);
         } 
     }
@@ -91,9 +92,11 @@ class RewardController extends BaseController
             ], [
                 'image.*.mimes' => 'The file should be in one of the formats: png, jpg, jpeg',
             ]);
-
-            Storage::disk('public')->delete($childReward->image->filename);
-            $childReward->image->delete();
+            
+            if($childReward->has('image')){
+                Storage::disk('public')->delete($childReward->image->filename);
+                $childReward->image->delete();
+            }
             
                 $path = $request->file('image')
                     ->store('reward-images', 'public');
